% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csd.R
\docType{data}
\name{csd}
\alias{csd}
\title{Data from the 'Critical Slowing Down'-Study.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1476 rows and 85 columns.}
\source{
\url{https://osf.io/j4fg8/}
}
\usage{
csd
}
\description{
a dataset of a single participant, with a history of major
     depressive disorder (MDD), whose daily life experience was monitored
     over the course of 239 days.
}
\details{
Data retrieved from single participant.

- Several columns were recoded from [-3, 3] scale to [1, 7] scale.

- dayno was recoded: the original data contained day number in the year (i.e.,
dayno == 1 for january 1st. The source codebook states that dayno codes "the day
number in the experiment". The variable was recalculated as the number of
days since the first assessment, to more accurately reflect this intend.
}
\examples{
# number of assessments
nrow(csd)

}
\references{
Kossakowski, J. J., Groot, P. C., Haslbeck, J. M. B., Borsboom, D., &
Wichers, M. (2017). Data from “Critical Slowing Down as a Personalized
Early Warning Signal for Depression.” Journal of Open Psychology Data, 5(1).
\url{https://doi.org/10.5334/jopd.29}

Wichers, M., & Groot, P. C. (2016). Critical Slowing Down as a
Personalized Early Warning Signal for Depression. Psychotherapy and
Psychosomatics. \url{https://doi.org/10.1159/000441458}
}
\keyword{datasets}
