% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_features_dataset.R
\name{generate_features_dataset}
\alias{generate_features_dataset}
\title{Generate time series used in "Feature Extraction"-chapter of the manual.}
\usage{
generate_features_dataset(seed = NULL, n_measurements_per_day = 5,
  n_days = 21, m = c(3, 6), sdev = c(0.5, 0.5), ampl = c(1, 4),
  range = c(0, 10), missingness_prob = c(0.1, 0.5),
  autocorrelation = 0.6, plot = FALSE)
}
\arguments{
\item{seed}{random seed.}

\item{n_measurements_per_day}{number of data points per day.}

\item{n_days}{assessment period (in days).}

\item{m}{mean at start (first element) and end (second element) of s.}

\item{sdev}{sd at start (first element) and end (second element) of s.}

\item{ampl}{amplitude of circadian rhythms (1-day, 7-day period) in s.}

\item{range}{min, max range of s.}

\item{missingness_prob}{minimal and maximal missingness probability of s.}

\item{autocorrelation}{strength of autocorrelation component in s.}

\item{plot}{if TRUE, the series is plotted (s ~ t). Default is FALSE.}
}
\value{
a data.frame, containing the generated timeseries, with time
in column t and the signal in column s.
}
\description{
Generate time series used in "Feature Extraction"-chapter of the manual.
}
\examples{
d <- generate_features_dataset(seed = 123, plot = TRUE)
head(d)

fm <- lm(s ~ t, d)
summary(fm)
}
