% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPSdata.R
\docType{data}
\name{GPSdata}
\alias{GPSdata}
\title{GPS data}
\format{A data frame with 587840 observations on the following 6 variables:
\describe{
  \item{\code{TimeStamp}}{Datetime of measurement}
  \item{\code{lat}}{GPS Latitude}
  \item{\code{lon}}{GPS Longitude}
  \item{\code{accuracy}}{GPS estimated accuracy (in meters)}
  \item{\code{activity}}{Activity prediction}
  \item{\code{activity_confidence}}{Altitude (in meters)}
  \item{\code{velocity}}{Velocity (in km per hour)}
  \item{\code{altitude}}{Altitude (in meters)}
}}
\source{
\url{jruwaard.nl}
}
\usage{
GPSdata
}
\description{
a Google timelie GPS dataset of a single participant.
}
\details{
Data retrieved from single participant.
}
\examples{
# number of assessments
nrow(GPSdata)

}
\keyword{datasets}
