% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plan.R
\name{sample_plan}
\alias{sample_plan}
\title{Generate a signal-contingent EMA sampling plan}
\usage{
sample_plan(n_participants = 7, n_days = 2, times = c("08:00-10:00",
  "12:30", "17:00-19:00"), plot = FALSE)
}
\arguments{
\item{n_participants}{number of participants in the study}

\item{n_days}{study period}

\item{times}{times (or time-periods) at which assessments are requested; a
character vector with time elements (e.g., "11:00", "23:00"), with as
many elements as n_beeps_per_day. Times can be specified as time points,
("11:15"), or as periods ("11:00-12:00"), in which case a random time
point is chosen, per participant, in the designated period.}

\item{plot}{if TRUE, the sample plan is plotted. Default is FALSE.}
}
\value{
a data.frame, containing the sample plan.
}
\description{
Generate a signal-contingent EMA sampling plan
}
\examples{
plan <- sample_plan(plot = TRUE)
head(plan)

}
