% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ema.R
\name{sim_ema}
\alias{sim_ema}
\title{Simulate EMA data}
\usage{
sim_ema(n_participants = 5, n_days = 7, times = c("08:00-10:00", "12:30",
  "17:00-19:00"), plot = FALSE, intercept = 0, intercept_var = 0,
  slope = 0, slope_var = 0, slope2 = 0, slope2_var = 0,
  error_var = 0.2, rho = 0.5)
}
\arguments{
\item{n_participants}{number of participants in the study}

\item{n_days}{study period}

\item{times}{times (or time-periods) at which assessments are requested; a
character vector with time elements (e.g., "11:00", "23:00"), with as
many elements as n_beeps_per_day. Times can be specified as time points,
("11:15"), or as periods ("11:00-12:00"), in which case a random time
point is chosen, per participant, in the designated period.}

\item{plot}{if TRUE, the sample plan is plotted. Default is FALSE.}

\item{intercept}{intercept}

\item{intercept_var}{variance of intercept, between individuals}

\item{slope}{slope}

\item{slope_var}{variance of slope, between individuals}

\item{slope2}{slope, squared}

\item{slope2_var}{variance of slope squared}

\item{error_var}{error variance}

\item{rho}{autocorrelation of errors}
}
\value{
a data.frame, containing the simulated data.
}
\description{
Simulate EMA data
}
\examples{
a <- sim_ema(n_participants = 20,
             intercept_var = 0.05,
             slope = -0.05,
             slope_var = 0.0005)
b <- sim_ema(n_participants = 20,
             intercept_var = 0.05,
             slope = -0.5,
             slope_var = 0.0005,
             slope2 = 0.02,
             slope2_var = 0)
b$id <- b$id + max(a$id)

c <- rbind(a, b)
c$group <- c(rep(0, nrow(a)), rep(1, nrow(b)))

library(ggplot2)
ggplot(data = c, aes(x = time, y = Y)) +
       geom_line(aes(group = id, colour = factor(group))) +
       geom_point(size = .8) +
       #geom_smooth(aes(group = factor(id), colour = factor(group)), method = "lm", se = FALSE) +
       #geom_smooth(aes(group = factor(group)), method = "lm", se = FALSE) +
       xlab("Time")

# test
library(nlme)
fm <- lme(Y ~ 1 + (time + I(time^2)) * group , random = ~ 1 + time | id,
         data = c,
         correlation = corAR1())
summary(fm)

}
