% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ema.R
\name{sim_ema}
\alias{sim_ema}
\title{Simulate EMA data}
\usage{
sim_ema(plan = sample_plan(), mm_par = list(fixed = c(intercept = 0,
  time = 0, time2 = 0), random = c(intercept = 0, time = 0, time2 = 0),
  error = 0.02, phi = 0), lim = NULL)
}
\arguments{
\item{plan}{a sample plan, as generated by \code{'sample_plan'}}

\item{mm_par}{a list, defining the parameters of a mixed effects model; the
list should contain elements "fixed", "random", "error" and "phi".}

\item{lim}{optional upper and lower limits of data (defaults to NULL, in
which case data values are unbounded).}
}
\value{
a \code{data.frame}, containing the simulated data:
     \item{id}{Participant id (numeric)}
     \item{observation}{observation counter (within participant; numeric)}
     \item{day}{day counter (numeric)}
     \item{beep}{beep counter, within day (numeric)}
     \item{t}{time of day (string, in "\%H:\%M" format)}
     \item{Y}{Simulated EMA value}
}
\description{
Simulate EMA data
}
\examples{
sample_plan <- sample_plan(n_participants = 20, n_days = 7)

mm_par_a <- list(
  fixed  = c(intercept = 0,
             time      = 0),
  random = c(intercept = 0.1,
             time      = 0.005),
  error  = 0.2,
  phi    = 0.5
)
a <- sim_ema(plan = sample_plan,
             mm_par = mm_par_a)

mm_par_b <- mm_par_a
mm_par_b$fixed['time']   <- -0.5
mm_par_b$random['time']  <- 0.005

b <- sim_ema(plan = sample_plan,
             mm_par = mm_par_b)

b$id <- b$id + max(a$id)

c <- rbind(a, b)
c$group <- c(rep(0, nrow(a)), rep(1, nrow(b)))

# plot
library(ggplot2)
ggplot(data = c, aes(x = time, y = Y)) +
       geom_line(aes(group = id, colour = factor(group))) +
       geom_point(size = .8) +
       geom_smooth(aes(group = factor(group), colour = factor(group)),
                   method = "loess", se = FALSE, size = 2) +
       xlab("Time")

# test
library(nlme)
fm <- lme(Y ~ 1 + time * group ,
         random = ~ 1 + time | id,
         data = c,
         correlation = corAR1())
summary(fm)


}
